// models/allocation.ts
export default interface IAllocation {
    id: number;
    taskId: string;
    description?: string;
    start?: string;
    end?: string;
    status?: string;
    assigned_to?: string;
    created_by?: string;
    end_date?: string;
    estimate?: string;
    start_date?: string;
    task_desc?: string;
    task_id?: string;
  }
  